apps/video-analysis-server/requirements.txt
# FastAPI Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0

# Authentication
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0

# Core AI/ML Dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
git+https://github.com/openai/CLIP.git

# Computer Vision
opencv-python>=4.8.0
mediapipe>=0.10.0
Pillow>=10.0.0

# Video Processing
scenedetect[opencv]>=0.6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Utilities
numpy>=1.24.0
aiofiles>=23.0.0
